+++ frontend/src/Components/BookCreator/BookCreator.jsx
@@ -1,68 +1,41 @@
 import React, { useState } from 'react';
-import './BookCreator.sass';
-import NavbarButtons from '../Shared/NavbarButtons/NavbarButtons';
-import SidebarLeftTopTop from '../Shared/SidebarLeftTopTop/SidebarLeftTopTop';
-import Carousel from '../Shared/NavBarCarrousel/NavBarCarrousel';
-import PageContainer from '../Shared/PageContainer/PageContainer';
-import Grids from '../Shared/Grids/Grids';
-import { FullScreen, useFullScreenHandle } from 'react-full-screen';
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import { faArrowsAlt, faExpandArrowsAlt } from '@fortawesome/free-solid-svg-icons';
+import './BookCreator.sass'
+import NavbarButtons from '../Shared/NavbarButtons/NavbarButtons'
+import SidebarLeftTopTop from '../Shared/SidebarLeftTopTop/SidebarLeftTopTop'
+import Carousel from '../Shared/NavBarCarrousel/NavBarCarrousel'
+import PageContainer from '../Shared/PageContainer/PageContainer'
+import { DndProvider } from 'react-dnd'
+import { HTML5Backend } from 'react-dnd-html5-backend'
+import { TouchBackend } from 'react-dnd-touch-backend'
+import { isMobile } from 'react-device-detect'
+
 
 const BookCreator = () => {
-  const handle = useFullScreenHandle();
-  const [buttonVisible, setButtonVisible] = useState(true);
 
-  const toggleButtonVisibility = (isVisible) => {
-    setButtonVisible(isVisible);
-  };
+  const backend = isMobile ? TouchBackend : HTML5Backend
 
-  const handleEnterFullScreen = () => {
-    toggleButtonVisibility(false);
-    handle.enter();
-  };
-
-  const handleExitFullScreen = () => {
-    toggleButtonVisibility(true);
-    handle.exit();
-  };
 
   return (
-    <>
-      <div className="container-fluid bookCreator">
-        <div className="row flex-nowrap contentBookCreator">
+    <DndProvider backend={backend}>
+      <div className='container-fluid bookCreator'>
+        <div className='row flex-nowrap contentBookCreator'>
           <SidebarLeftTopTop />
-          <div className="col-ms-10 p-0 mx-auto">
+          <div className='col-ms-10 p-0 mx-auto'>
             <NavbarButtons />
             <Carousel />
 
-            <FullScreen handle={handle}>
-              <div className="myactivity">
-                {!handle.active && (
-                  <div className="fullscreen-buttons">
-                    <button id="buttonExpand" onClick={handleEnterFullScreen}>
-                      <FontAwesomeIcon icon={faExpandArrowsAlt} />
-                      <i className="fa fa-expand"></i>
-                    </button>
-                  </div>
-                )}
+        
+   
+                
+
+            <PageContainer  title={'Activity 3'} />
 
          
-
-                <PageContainer
-                  title={'Activity 3'}
-                  image={'https://cdn.pixabay.com/photo/2018/01/14/23/12/nature-3082832_1280.jpg'}
-                  imageAlt={'landscape'}
-                  text={'Text'}
-                  
-                />
-              </div>
-            </FullScreen>
-          </div>
+        </div>
         </div>
       </div>
-    </>
-  );
-};
+    </DndProvider>
+  )
+}
 
-export default BookCreator;
+export default BookCreator
