+++ frontend/src/Components/BookCreator/BookCreator.sass
@@ -8,18 +8,7 @@
 .contentBookCreator
   height: auto
 
-#buttonExpand
-  margin-left: 81%
-  background-color: #3498db
-  color: #fff
-  padding: 10px 20px
-  border: none
-  border-radius: 5px
-  cursor: pointer
-  box-shadow: 2px 2px 5px rgba(0, 0, 0, 0.3)
 
-.fullscreen-buttons
-  background-color: white
 
-.myactivity
-  background-color: white
+
+
