+++ frontend/src/Components/Shared/PageContainer/PageContainer.jsx
@@ -8,6 +8,9 @@ import DoubleGridHorizontal from '../Grids/DoubleGridHorizontal/DoubleGridPlaceh
 import DoubleGridVertical from '../Grids/DoubleGridVertical/DoubleGridVerticalPlaceholder'
 import FullGrid from '../Grids/FullGrid/FullGridPlaceholder'
 import QuadrupleGrid from '../Grids/QuadrupleGrid/QuadrupleGridPlaceholder'
+import { FullScreen, useFullScreenHandle } from 'react-full-screen';
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
+import { faArrowsAlt, faExpandArrowsAlt } from '@fortawesome/free-solid-svg-icons';
 
 const widgetType = 'widgetType'
 
@@ -23,6 +26,7 @@ const widgetTypeToComponent = {
 }
 
 const PageContainer = ({ title }) => {
+  const [buttonVisible, setButtonVisible] = useState(true);
   const [droppedComponent, setDroppedComponent] = useState(null)
   const [saveData, setSaveData] = useState(null)
 
@@ -46,35 +50,60 @@ const PageContainer = ({ title }) => {
     setDroppedComponent(null)
   }
 
+  const handle = useFullScreenHandle();
+
+  const toggleButtonVisibility = (isVisible) => {
+    setButtonVisible(isVisible);
+  };
+
+  const handleEnterFullScreen = () => {
+    toggleButtonVisibility(false);
+    handle.enter();
+  };
+
   return (
     <div className='container-fluid'>
       <div className='row'>
         <div className='col'>
+        <FullScreen handle={handle}>
           <div className='card shadow mb-4 content_page'>
             <div className='card-header py-3 d-flex flex-row align-items-center justify-content-between'>
               <h6 className='m-0 font-weight-bold text-primary'>{title}</h6>
-  
-
-              <div className='card-body' width={'5000'}>
+              <div className='d-flex'>
                 <button onClick={remove}>
                   <img src='/escoba.png' alt='Clear' />
                 </button>
                 <button onClick={save}>
                   <img src='/expediente.png' alt='Save' />
                 </button>
+             
+                {!handle.active && (
+                  <div className="fullscreen-buttons">
+                    <button id="buttonExpand" onClick={handleEnterFullScreen}>
+                      <FontAwesomeIcon icon={faExpandArrowsAlt} />
+                      <i className="fa fa-expand"></i>
+                    </button>
+                  </div>
+                  
+                )}
+                 
               </div>
             </div>
-
+          
             <div
               className='card-body custom-card-body-page-container'
               ref={drop}
             >
+             
               {widgetTypeToComponent[droppedComponent] &&
                 React.createElement(widgetTypeToComponent[droppedComponent], {
                   saveData,
                 })}
             </div>
+
+         
           </div>
+          </FullScreen>
         </div>
       </div>
     </div>
