+++ frontend/src/Components/Shared/PageContainer/PageContainer.sass
@@ -1,12 +1,29 @@
 .content_page
   width: 85%
-  height: 100% !important
+  height: 800px !important
   margin-right: 90%
   margin-top: 1%
 
+#buttonExpand
+  background-color: #3498db
+  color: #fff
+  padding: 10px 20px
+  border: none
+  border-radius: 5px
+  cursor: pointer
+  box-shadow: 2px 2px 5px rgba(0, 0, 0, 0.3)
+
+@media screen and (display-mode: fullscreen) 
+  .content_page
+    width: 100% !important
+    height: 100% !important
+
+  
+
+
 @media (min-width: 1700px) and (max-width: 3000px)
   .content_page
-    width: 100%
+    width: 85%
     height: 78vh
     margin-right: 5%
 
@@ -76,13 +93,6 @@
     height: 33vh
     margin-right: 5%
 
-
-@media (min-height: 900px) and (max-height: 1200px)
-  .content_page
-    height: 80vh
-    margin-right: 5%
-
-
 @media (min-height: 800px) and (max-height: 1000px)
   .content_page
     height: 73vh
@@ -117,8 +127,4 @@
   .content_page
     height: 40vh
     margin-right: 5%
-
-
-
-
